extract IntUiBridgeDropDown
